Intermediate base class with Bookmark and Share
